Preliminary carrier calculations
Currently a bug exists on line 525 on mainCalc that it cannot check for
shipname for some reason. Action listener is always called. Right now
commented out but that means the calculate button will not be enabled if
slot 3 weapon is selected and nothing else.

Also carrier calculations is currently bugged and still outputs only 1
value.